Correctly encode/decode bindings like vm:sProperty:to
A binding such as `vm:sProperty:to` is encoded to `vm:s:u:property:to`, which is ambiguous to the decoder; should it decode `:s:` first or `:u:` first?

Previously, we decoded `:s:` first, which decoded `vm:s:u:property:to` to `vm u:property:to`. Now, we decode `:u:` first, which should be ok because only the new bindings need to encode uppercase with `:u:` and the new bindings don’t support spaces in them.

Fixes https://github.com/canjs/can-attribute-encoder/issues/11